test: split plugin loader coverage by concern
