+++ src/plugins/bundled-capability-metadata.test.ts
@@ -32,6 +32,7 @@ describe("bundled capability metadata", () => {
           manifest.contracts?.mediaUnderstandingProviders,
         ),
         imageGenerationProviderIds: uniqueStrings(manifest.contracts?.imageGenerationProviders),
+        webFetchProviderIds: uniqueStrings(manifest.contracts?.webFetchProviders),
         webSearchProviderIds: uniqueStrings(manifest.contracts?.webSearchProviders),
         toolNames: uniqueStrings(manifest.contracts?.tools),
       }))
@@ -42,6 +43,7 @@ describe("bundled capability metadata", () => {
           entry.speechProviderIds.length > 0 ||
           entry.mediaUnderstandingProviderIds.length > 0 ||
           entry.imageGenerationProviderIds.length > 0 ||
+          entry.webFetchProviderIds.length > 0 ||
           entry.webSearchProviderIds.length > 0 ||
           entry.toolNames.length > 0,
       )
